app/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
dataclasses-json==0.6.6
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
docker==7.1.0
docx2txt==0.8
exceptiongroup==1.2.1
executing==2.0.1
fastembed==0.2.7
filelock==3.14.0
flatbuffers==24.3.25
frontend==0.0.3
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.7
importlib-metadata==6.11.0
ipykernel==6.23.3
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.1
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langsmith==0.1.63
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
onnx==1.16.1
onnxruntime==1.18.0
openai==1.30.3
orjson==3.10.3
packaging==23.2
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
Pillow==9.5.0
platformdirs==4.2.2
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
Pygments==2.18.0
Pympler==1.0.1
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pyodbc==5.1.0
pypdf==4.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.2
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==2.7.0
sentencepiece==0.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlalchemy-iris==0.13.3
stack-data==0.6.3
starlette==0.37.2
streamlit==1.24.0
sympy==1.12
tenacity==8.3.0
testcontainers-core==0.0.1rc1
testcontainers-iris==1.1.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.2
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.39.3
triton==2.2.0
typing-inspect==0.9.0
typing_extensions==4.12.0
tzdata==2024.1
tzlocal==4.3.1
urllib3==2.2.1
uvicorn==0.29.0
validators==0.28.3
watchdog==4.0.1
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.2
